Initial Commit
This is a bash script that I created to initialize new web dev projects. It initializes the git repo, which will prompt you for the required info, runs npm init, which again prompts you for the required info, installs gulp, gulp-less, gulp-watch, browser-sync, bootstrap, and jQuery. I plan on adding a prompt system for other packages and libraries I commonly use, but these are the most basic and anything can be added to the script easily enough. It also creates the gulpfile and the basic index.html, index.js, and master.less files that you will need then opens it in vsCode and initializes browser sync, which starts a local server in the project's base directory and launches the default browser which confirms that everything works and reflects your changes as you make them.